Refactors EpisodeFetcher scripts
