+++ scripts/lib/publishing.rb
@@ -1,0 +1,2 @@
+module Publishing
+end
