+++ scripts/lib/publishing/service.rb
@@ -1,0 +1,4 @@
+module Publishing
+  module Service
+  end
+end
